refactor: 로또 한 개 생성하는 역할을 LottoFactory 에서 Lotto 로 옮김
